parse_wakeup_sources: support long names
The 12 character format string in the kernel is a minimum width, not a
maximum. The max size for a wakelock name is PAGE_SIZE, due to the limit
of the size of a sysfs attribute write. Accept names of at least 4096
bytes.

Add a line with a 4097-byte name to the test case.