wakeup_source: fix use after free by refcounting
Adding a destructor function for a wakeup_source in
53a02e33a3cda1b2fa8feb8ae7402c2032907cf9 introduced a use-after-free.
Fully dynamically allocate struct wakeup_source and refcount it to allow
the lifetime to be nonfixed.

This allow the client of parse_wakeup_sources to not be required to free
each wakeup_source, but allows it to extend the lifetime of the object
by referencing it.