handle interrupted system calls
This program does not set any non-default signal behavior, but Linux's
idiosyncratic behavior with SIGSTOP can still cause EINTR to be returned
for that signal. There is no way to configure that signal to be ignored
or automatically restart, making SA_RESTART useless as a method to avoid
the programmer having to explicity handle EINTR.

Use glibc's TEMP_FAILURE_RETRY macro for all interruptible system calls.