parse_wakeup_sources: always free on error
This was attempting to free the "name" matched by sscanf in the event of
a partial match. This should always free the wakeup source, and the
"name" field will be freed as well if the pointer is initialized.

Also NULL out the output parameter so we don't return a pointer to freed
memory in the case of an error.